requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
kafka-python
requests
pyspark